Add ACL paths for proxying
Given our service URIs, we must also expose the ACL paths so that the services may be reached from the outside world. The ACL paths must have their paths preserved.

The main shortcoming of this approach compared with our forthcoming ACL implementation of ConductR 1.2 is that we do not test the HTTP method.